requirements.txt
sentry-sdk>=2.0.0
psutil>=5.9.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
